Chris correcion errores

+++ SistemaDeComprasYVentas/Commands/EliminarCuentaCommand.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
 
 namespace SistemaDeComprasYVentas.Commands
@@ -14,25 +15,44 @@ namespace SistemaDeComprasYVentas.Commands
 	public class EliminarCuentaCommand : CommandBase
 	{
 		private UsuarioRequests requests;
+		private OutputMessages messages;
 		private ICommand NavigateBuscarProductosCommand { get; }
 
 		public EliminarCuentaCommand()
 		{
 			requests = new UsuarioRequests();
+			messages = new OutputMessages();
 			NavigateBuscarProductosCommand = new NavigateCommand< BuscarPublicacionesViewModel >( 
 												 NavigationServiceCreator.GetInstance().CreateBuscarNavigationService() );
 		}
 
 		public override void Execute( object parameter )
 		{
-			requests.DeleteUsuario( LoginSession.GetInstance().ClaveUsuario, LoginSession.GetInstance().AccessToken ).ContinueWith( Task => 
+			if( ConfirmElimination() )
 			{
-				if( Task.Exception == null )
+				requests.DeleteUsuario( LoginSession.GetInstance().ClaveUsuario, LoginSession.GetInstance().AccessToken ).ContinueWith( Task =>
 				{
-					LoginSession.GetInstance().Logout();
-					NavigateBuscarProductosCommand.Execute( this );
-				}
-			} );
+					if( Task.Exception == null )
+					{
+						LoginSession.GetInstance().Logout();
+						NavigateBuscarProductosCommand.Execute( this );
+					}
+				} );
+			}
+		}
+
+		private bool ConfirmElimination()
+		{
+			var request = MessageBox.Show( messages.MensajeConfirmarEliminarCuenta(), messages.ConfirmTitle(), 
+										   MessageBoxButton.OKCancel );
+			if( request == MessageBoxResult.OK )
+			{
+				return true;
+			}
+			else
+			{
+				return false;
+			}
 		}
 	}
 }
